requirements.txt
torch>=2.5.1
torchvision>=0.20.1
transformers>=4.42
scikit-learn>=1.7.2
huggingface-hub>=0.36
requests>=2.32.5
pyyaml>=6.0.3
pillow>=11.3
python-dotenv>=1.2.1
accelerate>=0.30
datasets>=4.4.2